refactor: 시민제보 layout 파일 구조 확립 #11
